Abstract fetch to hook
