+++ packages/js/src/dash/components/term-filter.js
@@ -1,9 +1,7 @@
-import { useCallback, useEffect, useRef, useState } from "@wordpress/element";
+import { useCallback, useState } from "@wordpress/element";
 import { __ } from "@wordpress/i18n";
 import { AutocompleteField, Spinner } from "@yoast/ui-library";
-import { debounce } from "lodash";
-import { FETCH_DELAY } from "../../shared-admin/constants";
-import { fetchJson } from "../util/fetch-json";
+import { useFetch } from "../util/use-fetch";
 
 /**
  * @type {import("../index").Taxonomy} Taxonomy
@@ -34,33 +32,13 @@ const transformTerm = ( term ) => ( { name: term.slug, label: term.name } );
  * @returns {JSX.Element} The element.
  */
 export const TermFilter = ( { idSuffix, taxonomy, selected, onChange } ) => {
-	const [ isPending, setIsPending ] = useState( false );
-	const [ error, setError ] = useState();
 	const [ query, setQuery ] = useState( "" );
-	const [ terms, setTerms ] = useState( [] );
-	/** @type {MutableRefObject<AbortController>} */
-	const controller = useRef();
-
-	// This needs to be wrapped including settings the state, because the debounce return messes with the timing/events.
-	const handleTermQuery = useCallback( debounce( ( ...args ) => {
-		fetchJson( ...args )
-			.then( ( result ) => {
-				setTerms( result.map( transformTerm ) );
-				setError( undefined ); // eslint-disable-line no-undefined
-			} )
-			.catch( ( e ) => {
-				// Ignore abort errors, because they are expected.
-				if ( e?.name !== "AbortError" ) {
-					setError( {
-						variant: "error",
-						message: __( "Something went wrong", "wordpress-seo" ),
-					} );
-				}
-			} )
-			.finally( () => {
-				setIsPending( false );
-			} );
-	}, FETCH_DELAY ), [] );
+	const { data: terms = [], error, isPending } = useFetch( {
+		dependencies: [ taxonomy.links.search, query ],
+		url: createQueryUrl( taxonomy.links.search, query ),
+		options: { headers: { "Content-Type": "application/json" } },
+		prepareData: ( result ) => result.map( transformTerm ),
+	} );
 
 	const handleChange = useCallback( ( value ) => {
 		if ( value === null ) {
@@ -74,18 +52,6 @@ export const TermFilter = ( { idSuffix, taxonomy, selected, onChange } ) => {
 		setQuery( event?.target?.value?.trim()?.toLowerCase() || "" );
 	}, [] );
 
-	useEffect( () => {
-		setIsPending( true );
-		controller.current?.abort();
-		controller.current = new AbortController();
-		handleTermQuery( createQueryUrl( taxonomy.links.search, query ), {
-			headers: { "Content-Type": "application/json" },
-			signal: controller.current.signal,
-		} );
-
-		return () => controller.current?.abort();
-	}, [ taxonomy.links.search, query, handleTermQuery ] );
-
 	return (
 		<AutocompleteField
 			id={ `term--${ idSuffix }` }
@@ -96,7 +62,10 @@ export const TermFilter = ( { idSuffix, taxonomy, selected, onChange } ) => {
 			onQueryChange={ handleQueryChange }
 			placeholder={ __( "All", "wordpress-seo" ) }
 			nullable={ true }
-			validation={ error }
+			validation={ error && {
+				variant: "error",
+				message: __( "Something went wrong", "wordpress-seo" ),
+			} }
 		>
 			{ isPending && (
 				<div className="yst-autocomplete__option">
